esports_smart_pipeline_project/requirements.txt
apache-flink==1.18
apache-flink-libraries==1.18
python-dotenv==1.0.0
